Move DbReaderOptions to config module
