+++ src/config.rs
@@ -669,6 +669,46 @@ impl Default for DbOptions {
     }
 }
 
+#[derive(Clone, Deserialize, Serialize)]
+pub struct DbReaderOptions {
+    /// How frequently to poll for new manifest files. Refreshing the manifest file allows readers
+    /// to detect newly compacted data.
+    pub manifest_poll_interval: Duration,
+
+    /// For readers that refresh their checkpoint, this specifies the lifetime to use for the
+    /// created checkpoint. The checkpoint’s expire time will be set to the current time plus
+    /// this value. If not specified, then the checkpoint will be created with no expiry, and
+    /// must be manually removed. This lifetime must always be greater than
+    /// manifest_poll_interval x 2
+    pub checkpoint_lifetime: Option<Duration>,
+
+    /// The max size of a single in-memory table used to buffer WAL entries
+    /// Defaults to 64MB
+    pub max_memtable_bytes: u64,
+
+    #[serde(skip)]
+    pub block_cache: Option<Arc<dyn DbCache>>,
+
+    /// The Clock to use for manifest polling and checkpoint lifetime maintenance.
+    ///
+    /// Default: the default clock uses the local system time on the machine
+    #[serde(skip)]
+    #[serde(default = "crate::config::default_clock")]
+    pub clock: Arc<dyn Clock + Send + Sync>,
+}
+
+impl Default for DbReaderOptions {
+    fn default() -> Self {
+        Self {
+            manifest_poll_interval: Duration::from_secs(10),
+            checkpoint_lifetime: None,
+            max_memtable_bytes: 4096,
+            block_cache: default_block_cache(),
+            clock: Arc::new(SystemClock::default()),
+        }
+    }
+}
+
 #[allow(unreachable_code)]
 fn default_block_cache() -> Option<Arc<dyn DbCache>> {
     #[cfg(feature = "moka")]
